It draws at least...
Solved the crash, it's somehow related to our bitmap. Maybe it needs to be in 256 colors for the Xbox to use it??

For now though, we also have some odd issue with the texture not displaying correctly. Just a big brown smudge. Also the vertex coordinates aren't quite right, the example was doing floating point stuff as if from the center of the screen.